refactor(dashboard): move timestamp helpers from TimestampPicker to utils
Move utcParts, utcDateFromParts, msFromNanos, utcTimeWithNanos, and
parseTimeWithNanos into timestamp.ts alongside the existing RFC 3339
helpers. TimestampPicker now imports them instead of defining inline.
